Add propper testing
