+++ src/index.js
@@ -1,62 +1,13 @@
-const util = require('util');
 const fse = require('fs-extra');
 const {glob} = require('glob');
 const core = require('@actions/core');
 const exec = require('@actions/exec');
 const github = require('@actions/github');
-const changelogParser = require('changelog-parser');
-const npmFetch = require('npm-registry-fetch');
 
-/**
- * @param {string} packageName 
- * @param {string} token
- */
-const fetchNPMVersions = async (packageName, token) => {
-    const versions = /** @type {Record<string, Version>} */(await npmFetch.json(
-        `http://registry.npmjs.org/${packageName}`,
-        {token},
-    ));
-
-    return Object.values(versions).map(({version}) => version);
-};
-
-const parseChangelog = util.promisify(changelogParser);
-
-/**
- * @param {string} changelogPath
- * @param {string} [notesPath]
- * @returns {Promise<string>}
- */
-const getBody = async (changelogPath, notesPath) => {
-    if (Boolean(notesPath)) {
-        core.info(`Notes Path found: ${notesPath}`);
-
-        try {
-            /** @type {string[]} */
-            const outputContent = [];
-            const filePaths = await glob(`${notesPath}/*.md`);
-
-            core.debug(JSON.stringify(filePaths));
-
-            await Promise.all(filePaths.map(async (filePath) => {
-                const fileContent = await fse.readFile(filePath, {encoding: 'utf-8'});
-                if (fileContent) {
-                    outputContent.push(fileContent.trim());
-                }
-            }));
-
-            core.debug(JSON.stringify(outputContent));
-
-            return outputContent.join('\n');
-        } catch (e) {
-            core.info('Failed Finding Release Notes');
-            core.error(e instanceof Error ? e.message : 'Unknown error');
-        }
-    }
-
-    const changelog = await parseChangelog(changelogPath);
-    return changelog.versions[0].body;
-};
+const {
+    getBody,
+    fetchNPMVersions
+} = require('./utils');
 
 async function run() {
     try {
@@ -163,11 +114,6 @@ async function run() {
 }
 
 run();
-
-/**
- * @typedef {object} Version
- * @prop {string} version
- */
 
 /**
  * @typedef {object} Package
