+++ src/tests/notes/bar.md
@@ -1,0 +1,1 @@
+This is bar
