+++ src/tests/notes/foo.md
@@ -1,0 +1,1 @@
+This is foo
